Revert "refactor query output functions out of gridParentComponents"
This reverts commit 51addcac76d2a21df150a8d95f54f061aab6ac7a.
