+++ src/sql/parts/grid/views/editData/editData.component.ts
@@ -22,7 +22,6 @@ import { GridParentComponent } from 'sql/parts/grid/views/gridParentComponent';
 import { EditDataGridActionProvider } from 'sql/parts/grid/views/editData/editDataGridActions';
 import { error } from 'sql/base/common/log';
 import { clone } from 'sql/base/common/objects';
-import * as GridContentEvents from 'sql/parts/grid/common/gridContentEvents';
 
 export const EDITDATA_SELECTOR: string = 'editdata-component';
 
@@ -111,18 +110,11 @@ export class EditDataComponent extends GridParentComponent implements OnInit, On
 			self._cd.detectChanges();
 		});
 
-		this.subscribeWithDispose(this.dataService.gridContentObserver, (type) => {
-			switch (type) {
-				case GridContentEvents.RefreshContents:
-					self.refreshResultsets();
-					break;
-				case GridContentEvents.ResizeContents:
-					self.resizeGrids();
-					break;
-			}
-		});
-
 		this.dataService.onAngularLoaded();
+	}
+
+	protected initShortcuts(shortcuts: { [name: string]: Function }): void {
+		// TODO add any Edit Data-specific shortcuts here
 	}
 
 	public ngOnDestroy(): void {
